production/feature_download/requirements.txt
sqlalchemy==1.4.39
binance-connector
psycopg2-binary